task 4, int bien puesto

+++ python_variable_annotations/4-define_variables.py
@@ -3,7 +3,7 @@
 '''variables con su tipo y contenido'''
 
 
-a: float = 1
+a: int = 1
 pi: float = 3.14
 i_understand_annotations: bool = True
 school: str = "Holbrton"
